Load parameters in main.tex instead of through \everyjob
The \everyjob approach was used because I thought it necessary to issue
a \special with the page dimensions for dvips. That, in turn, was
achieved at best with \everyjob, and it required that the parameters be
already loaded.

However, it turns out the \special is not actually necessary, so we can
get rid of this.